home commit: Background APIs done

+++ src/main/java/Controllers/Backgrounds.java
@@ -1,6 +1,7 @@
 package Controllers;
 
 import Server.Main;
+import com.sun.jersey.multipart.FormDataParam;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 
@@ -57,9 +58,19 @@ public class Backgrounds {
 
     }
 
-    public static void update(String backgroundPrice, String backgroundImage, int backgroundID) {
+    @POST
+    @Path("update")
+    @Consumes(MediaType.MULTIPART_FORM_DATA)
+    @Produces(MediaType.APPLICATION_JSON)
+    public String update(@FormDataParam("backgroundPrice") String backgroundPrice, @FormDataParam("backgroundImage") String backgroundImage, @FormDataParam("backgroundID") Integer  backgroundID) {
 
         try {
+            if (backgroundPrice == null || backgroundImage == null || backgroundImage == null) {
+                throw new Exception("One or more form data parameters are missing in the HTTP request.");
+            }
+
+            System.out.println("thing/update id=" + backgroundID);
+
             PreparedStatement ps = Main.db.prepareStatement("UPDATE backgrounds SET backgroundPrice = ?, backgroundImage = ? WHERE backgroundID = ?");
 
             /*The following code will update the entity with the ID specified from the backgrounds table to have the attributes that the user asks for*/
@@ -68,24 +79,40 @@ public class Backgrounds {
             ps.setInt(3, backgroundID);
 
             ps.executeUpdate();
+            return "{\"status\": \"OK\"}";
 
             /*If there is an error, the following code will identify this and will display an error message rather than crashing the program*/
         } catch (Exception exception) {
             System.out.println("Database error" + exception.getMessage());
+            return "{\"error\": \"Unable to update item, please see server console for more info.\"}";
         }
 
     }
 
-    public static void delete(int backgroundID) {
+    @POST
+    @Path("delete")
+    @Consumes(MediaType.MULTIPART_FORM_DATA)
+    @Produces(MediaType.APPLICATION_JSON)
+
+    public String delete(@FormDataParam("backgroundID") Integer backgroundID) {
 
         try {
+            if (backgroundID == null) {
+                throw new Exception("One or more form data parameters are missing in the HTTP request.");
+            }
+            System.out.println("thing/delete id=" + backgroundID);
+
             PreparedStatement ps = Main.db.prepareStatement("DELETE FROM backgrounds WHERE backgroundID = ?");
 
             ps.setInt(1, backgroundID);
 
             ps.executeUpdate();
+            return "{\"status\": \"OK\"}";
+
         } catch (Exception exception) {
             System.out.println("Database error" + exception.getMessage());
+            return "{\"error\": \"Unable to delete item, please see server console for more info.\"}";
+
         }
 
     }
